+++ examples/00_helloworld/00_helloworld_qt.cpp
@@ -4,7 +4,6 @@
 #include <hsmcpp/HsmEventDispatcherQt.hpp>
 #include <QCoreApplication>
 
-using namespace std::chrono_literals;
 using namespace hsmcpp;
 
 enum class States
@@ -26,13 +25,13 @@ int main(int argc, char** argv)
     hsm.registerState(States::OFF, [&hsm](const VariantVector_t& args)
     {
         printf("Off\n");
-        std::this_thread::sleep_for(1000ms);
+        std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });
     hsm.registerState(States::ON, [&hsm](const VariantVector_t& args)
     {
         printf("On\n");
-        std::this_thread::sleep_for(1000ms);
+        std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });