+++ examples/00_helloworld/00_helloworld_std.cpp
@@ -3,7 +3,6 @@
 #include <hsmcpp/hsm.hpp>
 #include <hsmcpp/HsmEventDispatcherSTD.hpp>
 
-using namespace std::chrono_literals;
 using namespace hsmcpp;
 
 enum class States
@@ -25,13 +24,13 @@ int main(const int argc, const char**argv)
     hsm.registerState(States::OFF, [&hsm](const VariantVector_t& args)
     {
         printf("Off\n");
-        std::this_thread::sleep_for(1000ms);
+        std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });
     hsm.registerState(States::ON, [&hsm](const VariantVector_t& args)
     {
         printf("On\n");
-        std::this_thread::sleep_for(1000ms);
+        std::this_thread::sleep_for(std::chrono::milliseconds(1000));
         hsm.transition(Events::SWITCH);
     });