+++ examples/05_timers/05_timers_generated.cpp
@@ -1,7 +1,6 @@
 #include "gen/TimerHsmBase.hpp"
 #include <hsmcpp/HsmEventDispatcherGLibmm.hpp>
 
-using namespace std::chrono_literals;
 using namespace hsmcpp;
 
 class TimerHsm: public TimerHsmBase