+++ tests/emptyapp.cpp
@@ -2,13 +2,11 @@
 #include <chrono>
 #include <thread>
 
-using namespace std::chrono_literals;
-
 int main(const int argc, const char**argv)
 {
     printf("Off\n");
-    std::this_thread::sleep_for(1000ms);
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
     printf("On\n");
-    std::this_thread::sleep_for(1000ms);
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
     return 0;
 }